requirements.txt
mysql-connector-python==8.0.25
numpy==1.21.1
pandas==1.3.0
pkg-resources==0.0.0
protobuf==3.17.3
python-dateutil==2.8.2
pytz==2021.1
six==1.16.0
